feat(sdds-acore/theme-builder): create only necessary dirs
