+++ theme-builder/plugin_attribute_builder/src/main/kotlin/com/sdds/plugin/attributebuilder/GenerateAttributesTask.kt
@@ -1,12 +1,15 @@
 package com.sdds.plugin.attributebuilder
 
-import com.sdds.plugin.attributebuilder.internal.factory.GeneratorFactory
+import com.sdds.plugin.attributebuilder.internal.data.AttributeData
 import com.sdds.plugin.attributebuilder.internal.factory.KtAttributeGeneratorFactory
 import com.sdds.plugin.attributebuilder.internal.factory.XmlAttributeGeneratorFactory
 import com.sdds.plugin.attributebuilder.internal.factory.XmlDocumentBuilderFactory
 import com.sdds.plugin.core.serializer.Serializer
 import com.sdds.plugin.core.serializer.decode
+import com.sdds.plugin.core.token.ColorToken
 import com.sdds.plugin.core.token.Theme
+import com.sdds.plugin.core.token.Token
+import com.sdds.plugin.core.token.isLight
 import com.sdds.plugin.core.utils.unsafeLazy
 import org.gradle.api.DefaultTask
 import org.gradle.api.file.DirectoryProperty
@@ -35,36 +38,40 @@ abstract class GenerateAttributesTask : DefaultTask() {
     abstract val target: Property<AttributeBuilderTarget>
 
     /**
-     * Директория для сохранения xml-файлов с атрибутами
+     * Директория проекта
      */
     @get:OutputDirectory
-    abstract val outputResDir: DirectoryProperty
+    abstract val projectDir: DirectoryProperty
 
     /**
-     * Директория для сохранения kt-файлов токенов
+     * Путь к директория для сохранения xml-файлов с атрибутами
      */
-    @get:OutputDirectory
-    abstract val outputDir: DirectoryProperty
+    @get:Input
+    abstract val outputResDirPath: Property<String>
+
+    /**
+     * Путь к директории для сохранения kt-файлов с атрибутами
+     */
+    @get:Input
+    abstract val outputDirPath: Property<String>
 
     private val xmlDocumentBuilderFactory by unsafeLazy { XmlDocumentBuilderFactory() }
 
     private val xmlAttributeGeneratorFactory by unsafeLazy {
-        XmlAttributeGeneratorFactory(xmlDocumentBuilderFactory, outputResDir.get().asFile)
+        XmlAttributeGeneratorFactory(xmlDocumentBuilderFactory)
     }
 
     private val ktAttributeGeneratorFactory by unsafeLazy {
         KtAttributeGeneratorFactory()
     }
 
-    private val generatorFactory by unsafeLazy {
-        GeneratorFactory(
-            target = target.get(),
-            xmlAttributeGeneratorFactory = xmlAttributeGeneratorFactory,
-            ktAttributeGeneratorFactory = ktAttributeGeneratorFactory,
-        )
+    private val xmlAttributeGenerator by unsafeLazy {
+        xmlAttributeGeneratorFactory.create(projectDir.get().dir(outputResDirPath).get().asFile)
     }
 
-    private val attributeGenerator by unsafeLazy { generatorFactory.createAttributeGenerator() }
+    private val ktAttributeGenerator by unsafeLazy {
+        ktAttributeGeneratorFactory.create()
+    }
 
     /**
      * Генерирует файлы с атрибутами
@@ -72,9 +79,31 @@ abstract class GenerateAttributesTask : DefaultTask() {
     @TaskAction
     fun generate() {
         val tokens = decodeTheme().tokens
-        attributeGenerator.generateFrom(tokens)
+        val attributeData = tokens.extractAttributeData()
+        when (val target = target.get()) {
+            is AttributeBuilderTarget.View -> xmlAttributeGenerator.generate(
+                attributeData = attributeData,
+                attrPrefix = target.attrPrefix,
+            )
+
+            is AttributeBuilderTarget.Compose -> ktAttributeGenerator.generate(
+                attributeData = attributeData,
+            )
+        }
     }
 
     private fun decodeTheme(): Theme =
         themeFile.get().asFile.decode<Theme>(decoder = Serializer.meta)
+
+    private fun List<Token>.extractAttributeData(): AttributeData {
+        val colors = mutableListOf<String>()
+
+        this.forEach {
+            when (it) {
+                is ColorToken -> if (it.isLight) colors.add(it.displayName)
+            }
+        }
+
+        return AttributeData(colors)
+    }
 }
